pyproject.toml
[project]
name = "202508-aws-genai-workshop-material"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.3.1",
    "pyarrow>=21.0.0",
    "weaviate-client>=4.16.7",
]
